Fix Mongo specs
This updates the Mongo API and fixes specs to get a green test suite.
However the MongoResource model is not production-ready as it has been
currently designed. It does not take advantage of chunks in the GridFS
to load file ranges and all file operations load the entire file into
memory.